+++ src/components/Card.jsx
@@ -1,53 +1,135 @@
 import React from 'react';
 import styled from 'styled-components';
+import { AiFillStar, AiOutlineStar } from 'react-icons/ai';
+import { RiArrowDownSLine } from 'react-icons/ri';
 
-export default function Card({ movie, handleCardClick, handleLikeClick }) {
+export default function MyListCard({
+  movie,
+  handleCardClick,
+  handleLikeClick,
+}) {
   return (
     <CardWrapper>
-      <Figure onClick={() => handleCardClick(movie.id)}>
+      <Figure>
         <Image src={movie.large_cover_image} alt={movie.title} />
       </Figure>
-      <Title>{movie.title}</Title>
-      <LikeBox onClick={() => handleLikeClick(movie.id, movie.like)}>
-        {movie.like ? <Like>💖</Like> : <Like>🤍</Like>}
-      </LikeBox>
+      <Content>
+        <Title>{movie.title}</Title>
+        <Description>{movie.description_full}</Description>
+        <Genres>
+          {movie.genres.map((genre, index) => (
+            <Genre key={index}>{genre}</Genre>
+          ))}
+        </Genres>
+        <CardButtonWrapper>
+          <CardButton onClick={() => handleLikeClick(movie.id, movie.like)}>
+            {movie.like ? <AiFillStar /> : <AiOutlineStar />}
+          </CardButton>
+          <CardButton onClick={() => handleCardClick(movie.id)}>
+            <RiArrowDownSLine />
+          </CardButton>
+        </CardButtonWrapper>
+      </Content>
     </CardWrapper>
   );
 }
 
 const CardWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  position: relative;
+  height: 400px;
+  color: white;
+  background-color: #181717;
+  border-radius: 5px;
   position: relative;
 `;
 
 const Figure = styled.figure`
+  flex-grow: 1;
   margin: 0;
   overflow: hidden;
+  border-radius: 5px;
 `;
 
 const Image = styled.img`
-  display: block;
   width: 100%;
+  height: 100%;
   transition: 200ms;
+  cursor: pointer;
 
   &:hover {
-    cursor: pointer;
     transform: scale(1.05);
   }
+`;
+
+const Content = styled.div`
+  position: absolute;
+  bottom: 0;
+  width: 100%;
+  background-color: #29292980;
+  padding: 1rem;
+  display: flex;
+  flex-direction: column;
 `;
 
 const Title = styled.p`
   margin: 0;
   margin-top: 6px;
   font-size: 14px;
-  color: gray;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
 `;
 
-const Like = styled.span`
-  font-size: 30px;
+const Description = styled.p`
+  font-size: 1rem;
+  text-align: center;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
 `;
-const LikeBox = styled.div`
-  position: absolute;
-  top: -15px;
-  right: -15px;
+
+const Genres = styled.ul`
+  font-size: 1rem;
+
+  margin: 0;
+  margin-bottom: 1rem;
+  padding: 0;
+  display: flex;
+  list-style: none;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+`;
+
+const Genre = styled.li`
+  margin-right: 0.5rem;
+  font-size: 13px;
+`;
+
+const CardButtonWrapper = styled.div`
+  display: flex;
+  justify-content: space-between;
+`;
+
+const CardButton = styled.button`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 3rem;
+  height: 3rem;
+  color: white;
+  background-color: #181717;
+  border: 1px solid gray;
+  border-radius: 50%;
   cursor: pointer;
+
+  &:hover {
+    border: 2px solid gray;
+  }
+
+  & > * {
+    font-size: 24px;
+  }
 `;
